build: allow sandboxed builds to succeed
Problem:
elm-codegen cannot be built as-is in a sandboxed environment
such as Nix.
The npm dependencies with no hash listed may be nondeterministic,
which is disallowed in reproducible builds.

There are workarounds, such as applying this patch in the build
process, but I'd like to upstream this patch to share it with other
elm-codegen users.

Solution:
Use jeslie0/npm-lockfile-fix to fill in the missing hashes.

https://github.com/jeslie0/npm-lockfile-fix

---

I can share the Nix derivation I'm using for elm-codegen, if there's
interest. It's currently in a private repo.